Move GetAllMembers() to already existing Bogus.Platform/ExtensionsForType.cs.
