+++ Source/Bogus/Binder.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Runtime.CompilerServices;
+using Bogus.Platform;
 
 namespace Bogus
 {
